requirements.txt
fastapi==0.95.1
uvicorn==0.22.0

SQLAlchemy==2.0.13
psycopg2==2.9.6
alembic==1.11.1

python-dotenv==1.0.0

pydantic[email]
passlib[bcrypt]
bcrypt==4.0.1
pre-commit==3.3.3

streamlit==1.30.0
streamlit-authenticator==0.3.1
streamlit-option-menu==0.3.12
python-dotenv==1.0.1 
streamlit-calendar==1.1.0 